Photon Vision Subsystem almost done
Figured out the distance calculation stuff got it within .10 of a meter with unprecise values in the calculation which is pretty good but ill add the actual values in once I know where the cameras will be on the robot. Also created the methods and doubles that will be called in the commands to get the distance to the goal/ball, set the alliance pipeline, and find the X displacement of the goal. :)